Moved around rocketpack code
