+++ pkg/rocketblend2/rocketblend.go
@@ -219,7 +219,6 @@ func (d *driver) ResolveBlendFile(ctx context.Context) (*blendfile.BlendFile, er
 }
 
 func (d *driver) getDependencies(ctx context.Context) (map[reference.Reference]*rocketpack.RocketPack, error) {
-	// TODO: make sure GetPackages returns dependencies from builds as separate packages
 	return d.rocketPackService.GetPackages(ctx, d.blendConfig.RocketFile.GetDependencies()...)
 }
